+++ src/Stack/Types/Config.hs
@@ -62,6 +62,8 @@ module Stack.Types.Config
   -- * Details
   -- ** ApplyGhcOptions
   ,ApplyGhcOptions(..)
+  -- ** CabalConfigKey
+  ,CabalConfigKey(..)
   -- ** ConfigException
   ,HpackExecutable(..)
   ,ConfigException(..)
@@ -321,6 +323,8 @@ data Config =
          -- ^ Additional GHC options to apply to specific packages.
          ,configGhcOptionsByCat     :: !(Map ApplyGhcOptions [Text])
          -- ^ Additional GHC options to apply to categories of packages
+         ,configCabalConfigOpts     :: !(Map CabalConfigKey [Text])
+         -- ^ Additional options to be passed to ./Setup.hs configure
          ,configSetupInfoLocations  :: ![SetupInfoLocation]
          -- ^ Additional SetupInfo (inline or remote) to use to find tools.
          ,configPvpBounds           :: !PvpBounds
@@ -371,6 +375,27 @@ configProjectRoot c =
     PCProject (_, fp) -> Just $ parent fp
     PCGlobalProject -> Nothing
     PCNoProject _deps -> Nothing
+
+-- | Which packages do configure opts apply to?
+data CabalConfigKey
+  = CCKTargets -- ^ See AGOTargets
+  | CCKLocals -- ^ See AGOLocals
+  | CCKEverything -- ^ See AGOEverything
+  | CCKPackage !PackageName -- ^ A specific package
+  deriving (Show, Read, Eq, Ord)
+instance FromJSON CabalConfigKey where
+  parseJSON = withText "CabalConfigKey" parseCabalConfigKey
+instance FromJSONKey CabalConfigKey where
+  fromJSONKey = FromJSONKeyTextParser parseCabalConfigKey
+
+parseCabalConfigKey :: Monad m => Text -> m CabalConfigKey
+parseCabalConfigKey "$targets" = pure CCKTargets
+parseCabalConfigKey "$locals" = pure CCKLocals
+parseCabalConfigKey "$everything" = pure CCKEverything
+parseCabalConfigKey name =
+  case parsePackageName $ T.unpack name of
+    Nothing -> fail $ "Invalid CabalConfigKey: " ++ show name
+    Just x -> pure $ CCKPackage x
 
 -- | Which packages do ghc-options on the command line apply to?
 data ApplyGhcOptions = AGOTargets -- ^ all local targets
@@ -733,6 +758,8 @@ data ConfigMonoid =
     -- ^ See 'configGhcOptionsAll'. Uses 'Monoid.Dual' so that options
     -- from the configs on the right come first, so that they can be
     -- overridden.
+    ,configMonoidCabalConfigOpts     :: !(MonoidMap CabalConfigKey (Monoid.Dual [Text]))
+    -- ^ See 'configCabalConfigOpts'.
     ,configMonoidExtraPath           :: ![Path Abs Dir]
     -- ^ Additional paths to search for executables in
     ,configMonoidSetupInfoLocations  :: ![SetupInfoLocation]
@@ -854,6 +881,9 @@ parseConfigMonoidObject rootDir obj = do
 
         configMonoidGhcOptionsByName = coerce $ Map.fromList
             [(name, opts) | (GOKPackage name, opts) <- Map.toList options]
+
+    configMonoidCabalConfigOpts' <- obj ..:? "configure-options" ..!= mempty
+    let configMonoidCabalConfigOpts = coerce (configMonoidCabalConfigOpts' :: Map CabalConfigKey [Text])
 
     configMonoidExtraPath <- obj ..:? configMonoidExtraPathName ..!= []
     configMonoidSetupInfoLocations <-